+++ src/core/utils/matrix.js
@@ -51,6 +51,16 @@ function identity( outMatrix ) {
 
 }
 
+function zero( outMatrix ) {
+
+	for ( let r = 0, tr = outMatrix.length; r < tr; r ++ ) {
+
+		outMatrix[ r ].fill( 0 );
+
+	}
+
+}
+
 function scale( outMatrix, matrix, scalar ) {
 
 	for ( let r = 0, tr = outMatrix.length; r < tr; r ++ ) {
@@ -169,13 +179,11 @@ function svd( ru, rq, rv, matrix ) {
 
 	}
 
-	const qrows = q.length;
-	for ( let r = 0; r < qrows; r ++ ) {
+	// Set singular values on diagonal (assumes rq is zeroed)
+	const qlen = q.length;
+	for ( let r = 0; r < qlen; r ++ ) {
 
-		const rqrow = rq[ r ];
-		const qval = q[ r ];
-		rqrow.fill( 0 );
-		rqrow[ r ] = qval;
+		rq[ r ][ r ] = q[ r ];
 
 	}
 
@@ -299,6 +307,7 @@ function set( matrix, r, c, value ) {
 export const mat = {
 	transpose,
 	identity,
+	zero,
 	scale,
 	multiply,
 	create,